Use shortened config name for Azure artifact name only if len > 80.
The old method of truncating the Azure artifact name to 80 characters
did not guarantee that it was unique for a given build, because it was
possible that the CONFIG name between all builds in the matrix was not
unique out to 57 characters (80 - len("conda_artifacts_NNNNNN_")). This
commit propagates short_config_name, which is hashed for uniqueness, to
the Azure SHORT_CONFIG variable and uses that for the artifact name when
shortening is required.